Make the methods of saving checkpoints configurable

+++ train.py
@@ -238,28 +238,40 @@ def train_and_evaluate(rank, epoch, hps, nets, optims, schedulers, scaler, loade
 
       if global_step % hps.train.eval_interval == 0:
         evaluate(hps, net_g, eval_loader, writer_eval)
-        checkpoint_num_to_keep = 4
-        checkpoint_num = utils.number_of_checkpoints(hps.model_dir, "G_*.pth")
+
+        should_auto_delete_old_checkpoints = False
+        if hasattr(hps, 'checkpoints'):
+          if hps.checkpoints.auto_delete_old_checkpoints:
+            should_auto_delete_old_checkpoints = True
+        else:
+          logger.info(("The 'checkpoints' config option hasn't been specified!"))
+
         g_checkpoint_path = os.path.join(hps.model_dir, "G_{}_{}.pth".format(hps.model_name, global_step))
         d_checkpoint_path = os.path.join(hps.model_dir, "D_{}_{}.pth".format(hps.model_name, global_step))
-        if checkpoint_num < checkpoint_num_to_keep: 
-          utils.save_checkpoint(net_g, optim_g, hps.train.learning_rate, epoch, g_checkpoint_path)
-          utils.save_checkpoint(net_d, optim_d, hps.train.learning_rate, epoch, d_checkpoint_path)
-        else:
-          # Only keep a number of latest checkpoints to save disc space
-          # First save the latest checkpoint into the oldest checkpoint file.
-          # Then, rename the newly saved file to latest checkpoint name.
-          g_oldest_checkpoint_path = utils.oldest_checkpoint_path(hps.model_dir, "G_*.pth")
-          utils.save_checkpoint(net_g, optim_g, hps.train.learning_rate, epoch, g_oldest_checkpoint_path, False)
-          logger.info("Saving model and optimizer state at iteration {} to {}".format(
-    epoch, g_checkpoint_path))
-          os.rename(g_oldest_checkpoint_path, g_checkpoint_path)
+        utils.save_checkpoint(net_g, optim_g, hps.train.learning_rate, epoch, g_checkpoint_path)
+        utils.save_checkpoint(net_d, optim_d, hps.train.learning_rate, epoch, d_checkpoint_path)
 
-          d_oldest_checkpoint_path = utils.oldest_checkpoint_path(hps.model_dir, "D_*.pth")
-          logger.info("Saving model and optimizer state at iteration {} to {}".format(
-    epoch, d_checkpoint_path))
-          utils.save_checkpoint(net_d, optim_d, hps.train.learning_rate, epoch, d_oldest_checkpoint_path, False)
-          os.rename(d_oldest_checkpoint_path, d_checkpoint_path)
+        # Only keep a number of latest checkpoints to save disc space
+        if should_auto_delete_old_checkpoints:
+          num_checkpoints = utils.number_of_checkpoints(hps.model_dir, "G_*.pth")
+          num_checkpoints_to_keep = hps.checkpoints.num_checkpoints_to_keep
+          if hps.checkpoints.replace_old_checkpoints_mode:
+            if num_checkpoints >= num_checkpoints_to_keep:
+              # First save the latest checkpoint into the oldest checkpoint file.
+              # Then, rename the newly saved file to latest checkpoint name.
+              g_oldest_checkpoint_path = utils.oldest_checkpoint_path(hps.model_dir, "G_*.pth")
+              utils.save_checkpoint(net_g, optim_g, hps.train.learning_rate, epoch, g_oldest_checkpoint_path, False)
+              logger.info("Saving model and optimizer state at iteration {} to {}".format(epoch, g_checkpoint_path))
+              os.rename(g_oldest_checkpoint_path, g_checkpoint_path)
+
+              d_oldest_checkpoint_path = utils.oldest_checkpoint_path(hps.model_dir, "D_*.pth")
+              logger.info("Saving model and optimizer state at iteration {} to {}".format(epoch, d_checkpoint_path))
+              utils.save_checkpoint(net_d, optim_d, hps.train.learning_rate, epoch, d_oldest_checkpoint_path, False)
+              os.rename(d_oldest_checkpoint_path, d_checkpoint_path)
+            else:
+              if num_checkpoints > num_checkpoints_to_keep:
+                os.remove(utils.oldest_checkpoint_path(hps.model_dir, "G_*.pth"))
+                os.remove(utils.oldest_checkpoint_path(hps.model_dir, "D_*.pth"))      
     global_step += 1
   
   if rank == 0:
